blog/requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
exceptiongroup==1.2.1
Faker==26.0.0
h11==0.14.0
idna==3.7
lxml==5.2.2
outcome==1.3.0.post0
packaging==24.1
parameterized==0.9.0
pillow==10.3.0
psycopg2-binary==2.9.9
PySocks==1.7.1
python-dateutil==2.9.0.
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
urllib3==2.2.1
wcwidth==0.2.13
webdriver-manager==4.0.1
websocket-client==1.8.0
wsproto==1.2.0

